# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HiveMind Nexus API - Autonomous AI Workforce Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  idle: 'idle',
  thinking: 'thinking',
  working: 'working',
  meeting: 'meeting',
  offline: 'offline',
} as const;

export interface Agent {
  id: number;
  name: string;
  role: string;
  personality: string;
  status: AgentStatus;
  /** @nullable */
  currentTask?: string | null;
  productivityScore: number;
  avatarColor: string;
  model?: string;
  /** @nullable */
  lastActive?: string | null;
}

export interface NetworkNode {
  id: string;
  agentId: number;
  label: string;
  status: string;
  communicationCount?: number;
}

export interface NetworkEdge {
  source: string;
  target: string;
  weight: number;
  active?: boolean;
}

export interface AgentNetwork {
  nodes: NetworkNode[];
  edges: NetworkEdge[];
}

export type GoalStatus = typeof GoalStatus[keyof typeof GoalStatus];


export const GoalStatus = {
  pending: 'pending',
  active: 'active',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Goal {
  id: number;
  /** @nullable */
  userId?: string | null;
  title: string;
  description: string;
  status: GoalStatus;
  /** @nullable */
  industry?: string | null;
  createdAt: string;
}

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  industry?: string;
}

export type MessageMessageType = typeof MessageMessageType[keyof typeof MessageMessageType];


export const MessageMessageType = {
  statement: 'statement',
  question: 'question',
  disagreement: 'disagreement',
  approval: 'approval',
  task_handoff: 'task_handoff',
  user: 'user',
} as const;

export interface Message {
  id: number;
  goalId: number;
  /** @nullable */
  agentId: number | null;
  agentName: string;
  agentRole: string;
  content: string;
  messageType: MessageMessageType;
  /** @nullable */
  confidence?: number | null;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type MeetingStatus = typeof MeetingStatus[keyof typeof MeetingStatus];


export const MeetingStatus = {
  scheduled: 'scheduled',
  active: 'active',
  completed: 'completed',
} as const;

export interface Meeting {
  id: number;
  /** @nullable */
  goalId?: number | null;
  title: string;
  /** @nullable */
  agenda?: string | null;
  status: MeetingStatus;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  goalId?: number | null;
  agenda?: string;
}

export interface MeetingMessage {
  id: number;
  meetingId: number;
  agentId: number;
  agentName: string;
  agentRole: string;
  content: string;
  /** @nullable */
  confidence?: number | null;
  isDecision?: boolean;
  createdAt: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  blocked: 'blocked',
  failed: 'failed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Task {
  id: number;
  /** @nullable */
  goalId?: number | null;
  /** @nullable */
  assignedAgentId?: number | null;
  /** @nullable */
  assignedAgentName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface TaskInput {
  /** @nullable */
  goalId?: number | null;
  /** @nullable */
  assignedAgentId?: number | null;
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  blocked: 'blocked',
  failed: 'failed',
} as const;

export interface TaskUpdate {
  status?: TaskUpdateStatus;
  /** @nullable */
  assignedAgentId?: number | null;
  description?: string;
}

/**
 * @nullable
 */
export type BusinessMetricTrend = typeof BusinessMetricTrend[keyof typeof BusinessMetricTrend] | null;


export const BusinessMetricTrend = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export interface BusinessMetric {
  id: number;
  /** @nullable */
  goalId?: number | null;
  metricType: string;
  value: number;
  unit: string;
  label?: string;
  /** @nullable */
  trend?: BusinessMetricTrend;
  createdAt: string;
}

export type SimulationStatus = typeof SimulationStatus[keyof typeof SimulationStatus];


export const SimulationStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Simulation {
  id: number;
  goalId: number;
  timeframe: string;
  status: SimulationStatus;
  /** @nullable */
  predictions?: string | null;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export interface SimulationInput {
  goalId: number;
  timeframe: string;
}

export interface Report {
  id: number;
  /** @nullable */
  goalId?: number | null;
  title: string;
  content?: string;
  generatedBy: string;
  reportType?: string;
  createdAt: string;
}

export type GithubLogStatus = typeof GithubLogStatus[keyof typeof GithubLogStatus];


export const GithubLogStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface GithubLog {
  id: number;
  /** @nullable */
  commitHash?: string | null;
  commitMessage?: string;
  status: GithubLogStatus;
  /** @nullable */
  filesChanged?: number | null;
  /** @nullable */
  repoUrl?: string | null;
  createdAt: string;
}

export interface GithubPushInput {
  /** @minLength 1 */
  commitMessage: string;
}

export type GeneratedAssetAssetType = typeof GeneratedAssetAssetType[keyof typeof GeneratedAssetAssetType];


export const GeneratedAssetAssetType = {
  startup_name: 'startup_name',
  logo_concept: 'logo_concept',
  tagline: 'tagline',
  landing_page_copy: 'landing_page_copy',
  marketing_slogan: 'marketing_slogan',
  investor_pitch: 'investor_pitch',
  pricing_strategy: 'pricing_strategy',
  business_roadmap: 'business_roadmap',
} as const;

export interface GeneratedAsset {
  id: number;
  goalId: number;
  assetType: GeneratedAssetAssetType;
  title: string;
  content: string;
  generatedBy?: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalGoals: number;
  activeAgents: number;
  completedTasks: number;
  totalMessages: number;
  runningMeetings: number;
  autonomousMode?: boolean;
  companyHealth?: number;
  recentMetrics?: BusinessMetric[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  message: 'message',
  task: 'task',
  meeting: 'meeting',
  goal: 'goal',
  simulation: 'simulation',
  github: 'github',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  agentName: string;
  agentRole?: string;
  createdAt: string;
}

export type ListTasksParams = {
/**
 * @nullable
 */
goalId?: number | null;
/**
 * @nullable
 */
status?: string | null;
};

export type ListMetricsParams = {
/**
 * @nullable
 */
goalId?: number | null;
};