# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HiveMind Nexus API - Autonomous AI Workforce Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List business goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new business goal and trigger AI agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
  /goals/{id}/messages:
    get:
      operationId: listGoalMessages
      tags: [goals]
      summary: Get live agent conversation for a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendGoalMessage
      tags: [goals]
      summary: Send a prompt or inject a message into agent conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /goals/{id}/assets:
    get:
      operationId: listGoalAssets
      tags: [goals]
      summary: Get generated business assets for a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of generated assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedAsset"
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List all board meetings
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Start a new AI board meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Meeting started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get meeting details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
  /meetings/{id}/messages:
    get:
      operationId: listMeetingMessages
      tags: [meetings]
      summary: Get meeting discussion transcript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingMessage"
...
  /github/push:
    post:
      operationId: pushToGithub
      tags: [github]
      summary: Push project files to GitHub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPushInput"
      responses:
        "200":
          description: Push result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubLog"
...
components:
  schemas:
...
    Goal:
      type: object
      required: [id, title, description, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, active, completed, failed]
        industry:
          type: ["string", "null"]
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        industry:
          type: string
    Message:
      type: object
      required: [id, goalId, agentId, agentName, agentRole, content, messageType, createdAt]
      properties:
        id:
          type: integer
        goalId:
          type: integer
        agentId:
          type: ["integer", "null"]
        agentName:
          type: string
        agentRole:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [statement, question, disagreement, approval, task_handoff, user]
        confidence:
          type: ["number", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    Meeting:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        goalId:
          type: ["integer", "null"]
        title:
          type: string
        agenda:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, active, completed]
        createdAt:
          type: string
        endedAt:
          type: ["string", "null"]
    MeetingInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        goalId:
          type: ["integer", "null"]
        agenda:
          type: string
    MeetingMessage:
      type: object
      required: [id, meetingId, agentId, agentName, agentRole, content, createdAt]
      properties:
        id:
          type: integer
        meetingId:
          type: integer
        agentId:
          type: integer
        agentName:
          type: string
        agentRole:
          type: string
        content:
          type: string
        confidence:
          type: ["number", "null"]
        isDecision:
          type: boolean
        createdAt:
          type: string
...
    GithubLog:
      type: object
      required: [id, status, createdAt]
      properties:
        id:
          type: integer
        commitHash:
          type: ["string", "null"]
        commitMessage:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        filesChanged:
          type: ["integer", "null"]
        repoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    GithubPushInput:
      type: object
      required: [commitMessage]
      properties:
        commitMessage:
          type: string
          minLength: 1
    GeneratedAsset:
      type: object
      required: [id, goalId, assetType, title, content, createdAt]
      properties:
        id:
          type: integer
        goalId:
          type: integer
        assetType:
          type: string
          enum: [startup_name, logo_concept, tagline, landing_page_copy, marketing_slogan, investor_pitch, pricing_strategy, business_roadmap]
        title:
          type: string
        content:
          type: string
        generatedBy:
          type: string
        createdAt:
          type: string
...