# lib/api-zod/src/generated/types/generatedAsset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HiveMind Nexus API - Autonomous AI Workforce Platform
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedAssetAssetType } from './generatedAssetAssetType';

export interface GeneratedAsset {
  id: number;
  goalId: number;
  assetType: GeneratedAssetAssetType;
  title: string;
  content: string;
  generatedBy?: string;
  createdAt: string;
}