# lib/api-zod/src/generated/types/githubLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HiveMind Nexus API - Autonomous AI Workforce Platform
 * OpenAPI spec version: 0.1.0
 */
import type { GithubLogStatus } from './githubLogStatus';

export interface GithubLog {
  id: number;
  /** @nullable */
  commitHash?: string | null;
  commitMessage?: string;
  status: GithubLogStatus;
  /** @nullable */
  filesChanged?: number | null;
  /** @nullable */
  repoUrl?: string | null;
  createdAt: string;
}